requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
prophet
matplotlib
seaborn
requests